OD-11314 final adjustment after branch jdk18 was merged to trunk
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@49448 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ checkout-api/src/main/groovy/ish/oncourse/willow/functions/IsParentRequired.groovy
@@ -22,7 +22,7 @@ class IsParentRequired {
 
     boolean get() {
         if (contact.dateOfBirth) {
-            Integer age = ChronoUnit.YEARS.between(contact.dateOfBirth.toInstant(), new Date().toInstant())
+            Long age = ChronoUnit.YEARS.between(contact.dateOfBirth.toInstant(), new Date().toInstant())
             return new IsCollectParentDetails(college, context).get() && new GetContactAgeWhenNeedParent(college, context).get() > age
         }
         return false
